fix build

+++ AD5620.h
@@ -68,8 +68,8 @@ protected:
   bool     _hwSPI       = false;
   uint32_t _SPIspeed    = 16000000;
 
-  void     updateDevice(uint32_t value);
-  void     swSPI_transfer(uint32_t value);
+  void     updateDevice(uint16_t value);
+  void     swSPI_transfer(uint16_t value);
 
   __SPI_CLASS__ * _mySPI;
   SPISettings   _spi_settings;
